requirements.txt
gradio
requests
langgraph==0.3.21
langchain-google-genai==2.1.2
langgraph-prebuilt==0.1.7
google-genai==1.7.0
langchain
langchain-community
langchain-openai
langchain_huggingface
langchain-tavily
wikipedia
openai
unstructured
openpyxl